% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment}
\alias{experiment}
\title{Creates Instance of Experiment Object}
\usage{
experiment()
}
\value{
An Experiment object
}
\description{
Creates Instance of Experiment Object
}
\examples{
exp <- experiment()
}
